requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
attrs==25.1.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
distro==1.9.0
fastapi==0.115.8
fastapi-limiter==0.1.6
filetype==1.2.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0
greenlet==3.1.1
groq==0.18.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-anthropic==0.3.7
langchain-core==0.3.35
langchain-google-genai==2.0.9
langchain-groq==0.2.4
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langsmith==0.3.8
limits==4.0.1
multidict==6.1.0
numpy==1.26.4
openai==1.63.2
orjson==3.10.15
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pyparsing==3.2.1
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-mock==3.14.0
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.3.0b5
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
wrapt==1.17.2
yarl==1.18.3
zstandard==0.23.0
